refac: divide RoundStartBackdropContainer
